Move saving files to the socket server exclusively
